MM-31876: Fix racy test GetReplica
After starting a db connection, we would actually compare
the dbConn struct using reflection. This would race
with the connection cleaner goroutine which would modify
the internal structure of the db conn object.

There is no direct way to compare database connections safely
while it is being used. A hack might be to use the inveitable time.Sleep.
But it's cleaner to just remove the code altogether. It's only
testing a small part of the logic of assigning masters and replicas.

https://mattermost.atlassian.net/browse/MM-31876